+++ mex/common/transform.py
@@ -104,7 +104,7 @@ def split_to_camel(string: str) -> str:
 @lru_cache(maxsize=1024)
 def split_to_caps(string: str) -> str:
     """Convert the given string from `Split case` into `CAPS_CASE`."""
-    return "_".join(word.upper() for word in re.split("[^a-zA-Z]", string) if word)
+    return "_".join(word.upper() for word in re.split("[^a-zA-Z0-9]", string) if word)
 
 
 @lru_cache(maxsize=1024)