+++ mex/common/types/vocabulary.py
@@ -168,6 +168,12 @@ class BibliographicResourceType(VocabularyEnum):
     __vocabulary__ = "bibliographic-resource-type"
 
 
+class CodingSystem(VocabularyEnum):
+    """The type of a coding system."""
+
+    __vocabulary__ = "coding-system"
+
+
 class ConsentStatus(VocabularyEnum):
     """The status of a consent."""
 
@@ -178,6 +184,12 @@ class ConsentType(VocabularyEnum):
     """The type of a consent."""
 
     __vocabulary__ = "consent-type"
+
+
+class Country(VocabularyEnum):
+    """The type of a country."""
+
+    __vocabulary__ = "country"
 
 
 class DataProcessingState(VocabularyEnum):