+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - load langdetect profiles lazily to save ~150ms on every initial mex-common import
+- BREAKING: update mex-model depedency to v5.1
 
 ### Deprecated